Refactor rendering to be backend-agnostic.

+++ Modules/Orama.Rendering/RenderingModule.cs
@@ -5,6 +5,7 @@ using System.Runtime.InteropServices;
 
 using Orama.Common;
 using Orama.Math;
+using Orama.RHI;
 using Orama.Rendering.Components;
 using Orama.Rendering.Pipelines;
 using Orama.Rendering.Pipelines.Forward;
@@ -75,5 +76,5 @@ public class RenderingModule : BaseModule
 	/// <param name="renderable">The object to render.</param>
 	public void QueueObject(IClientRenderable renderable) => Renderables.Add(renderable);
 
-	public void OnResize(int width, int height) => Renderer.Resize(width, height);
+	public void OnResize(int width, int height) => Renderer.Resize((uint)width, (uint)height);
 }
